Create app.py (Copied From PyCharm)
This is the initial commit for this project. It utilizes an ML model throughout the whole process of utilizing the neurons, providing a better, alternate solution towards steps one and two (movement and neural behavior changes). This will still be buggy but it is a chance for me to get an initial commit down to show the progress that I have made so far.